track meta-swupdate krogoth branch
swupdate_git.bb used to always build the latest swupdate (as whatever
happens to be in swupdate master at the time of the build). For releases
we want the version we tested though. The release branches of meta-swupdate
now checkout a fixed version. Since we used to track meta-swupdate master,
rebase on kergoth since it is closest to master.