pupulate sstate-cached from master
Since there several machines now, it is nice if testing a patch against
all of them doesn't take a lot of time. This speeds up things a bit by
letting master populate the sstate-cache. The non master branches can
reuse this again.

The master branch itself does not use the sstate, but rebuilds it in a
temporary location, and moves it to the common place when the build was
succesfull.

While at it, the ci likes to upload and download artifacts between
stages. Since only the swu files are build for the non master branch,
just skip the artifacts passing completely.