ci: swap the sstate directories
The master branch builds sstate in a temporarily location and moves it
to the shared one once complete. If the directory already exists though
mv will make a subdirectory for it instead of replacing it. So move the
the target directory and delete it after the new one is available.