ci: use sstate for building debian packages
Since generating packages from Debian is now also based on rocko it is possible
to reuse the sstate from the regular OE build, so all native tools can't be reused
instead of build again. This does use the sstate from the last build instead of
the current one, but since the native tools seldomly change I do not think that
matters.