add a script to update recipes
Newer OE version get worse at handling tags, instead of looking them up
locally they bibake might reach out to the remote to fetch them even at
really weird time, during parsing e.g. To avoid that, this script fetches
the SRCREV and stores it in the recipe along with the package version.

In batch mode, the script removes the tag= from the giturl and adds the
corresponding UPSTREAM_CHECK_GITTAGREGEX / PV etc, by default to stdout,
when -s is given it will be stored as ${PN}_git.bb. If -g is also given
it will be staged for commiting.

In interactive mode it accepts changelog / commit like message and tries
to update the corresponding recipe and ask to commit it.